Fb format code

+++ contracts/modules/BurnModule.sol
@@ -7,5 +7,5 @@ import "../../openzeppelin-contracts-upgradeable/contracts/proxy/utils/Initializ
 abstract contract BurnModule is Initializable {
     bytes32 public constant BURNER_ROLE = keccak256("BURNER_ROLE");
 
-    event Burn (address indexed owner, uint amount);
+    event Burn(address indexed owner, uint256 amount);
 }
